refactor(ale_claw): rename harness/inference/ -> harness/model/
Plainer name for the model-call subpackage (model_config, unified_loop,
cache_policy, thinking, helper_runtime, image_sanitization, _message_shapes).
Pure rename: git-tracked moves, all import paths + test patch targets +
README layout updated. No logic change. 928/3 green.
