+++ ale_run/agents/ale_claw/harness/__init__.py
@@ -28,7 +28,7 @@ from .adapters import (
 )
 from .agent_loop import OpenClawComputerAgent
 from .tools.analyze_image import AnalyzeImageTool
-from .inference.cache_policy import (
+from .model.cache_policy import (
     OPENCLAW_CACHE_BOUNDARY,
     apply_openclaw_cache_markers,
     supports_anthropic_cache,
@@ -94,7 +94,7 @@ from .subagent.subagent_registry import (
     SubagentUsage,
 )
 from .subagent.subagent_tools import DelegateGeneralTool, DelegateGUITool, SubagentsTool
-from .inference.thinking import ThinkingConfig, ThinkLevel, resolve_thinking_default
+from .model.thinking import ThinkingConfig, ThinkLevel, resolve_thinking_default
 from .tools.tools import ToolLoggingCallback, build_tools, get_tool_summaries
 from .tools.tools_fs import EditFileTool, ReadFileTool, WriteFileTool
 from .tools.tools_shell import ExecTool
@@ -106,7 +106,7 @@ from .context.transcript import group_step_output
 # so sparse-checkout consumers that only pull the openclaw subpackage
 # still get the chat-completions OpenRouter route instead of falling
 # through to loops/openai.py (Responses API).
-from .inference import unified_loop  # noqa: F401
+from .model import unified_loop  # noqa: F401
 
 __all__ = [
     "OpenClawComputerAgent",
