+++ ale_run/agents/ale_claw/harness/agent_loop_helpers.py
@@ -12,8 +12,8 @@ import base64 as _base64
 import re
 from typing import TYPE_CHECKING, Any, Dict, List
 
-from .inference._message_shapes import _image_url_block
-from .inference.image_sanitization import (
+from .model._message_shapes import _image_url_block
+from .model.image_sanitization import (
     DEFAULT_LIMITS as _IMAGE_DEFAULT_LIMITS,
     sanitize_raw_image_bytes as _sanitize_raw_image_bytes,
 )
