+++ ale_run/agents/ale_claw/harness/model/image_sanitization.py
