+++ ale_run/agents/ale_claw/harness/model/thinking.py
