+++ tests/ale_claw/test_canonical_format.py
@@ -14,7 +14,7 @@ Covers:
 
 import json
 
-from ale_run.agents.ale_claw.harness.inference.model_config import resolve_model
+from ale_run.agents.ale_claw.harness.model.model_config import resolve_model
 
 from ale_run.agents.ale_claw.harness.canonical.canonical import (
     COMPACTION_PREAMBLE,
