+++ tests/ale_claw/test_openclaw_memory_flush.py
@@ -3,7 +3,7 @@
 import asyncio
 from unittest.mock import AsyncMock, MagicMock, patch
 
-from ale_run.agents.ale_claw.harness.inference.model_config import HelperTransportDefaults, ModelConfig, register_model_config
+from ale_run.agents.ale_claw.harness.model.model_config import HelperTransportDefaults, ModelConfig, register_model_config
 from ale_run.agents.ale_claw.harness.memory.memory_flush import run_memory_flush
 
 
@@ -58,12 +58,12 @@ class _MemoryStoreStub:
 
 class TestRunMemoryFlush:
     def setup_method(self):
-        from ale_run.agents.ale_claw.harness.inference.model_config import _MODEL_CONFIGS
+        from ale_run.agents.ale_claw.harness.model.model_config import _MODEL_CONFIGS
 
         self._original = list(_MODEL_CONFIGS)
 
     def teardown_method(self):
-        from ale_run.agents.ale_claw.harness.inference.model_config import _MODEL_CONFIGS
+        from ale_run.agents.ale_claw.harness.model.model_config import _MODEL_CONFIGS
 
         _MODEL_CONFIGS.clear()
         _MODEL_CONFIGS.extend(self._original)
